parisc: fix delay slots in the conditional instructions
The polarity of "need_nop" was in fact correct and the commit
7ca91a8cf43ba32765cd261b591562fc8ba9429b broke it.

The problem was that when a function contains a conditional jump to the
immediatelly following instruction, cgen_cond_jmp would evaluate it as a
forward branch (and omit the delay slot), while the CPU would evaluate it
as a backward branch and apply the delay slot logic for backward
branches.

Fix this by rejecting branches to the following instruction in
resolve_relocation and fall back to JMP_SHORT which doesn't have this
problem.