task: fix a race when spawning new thread
If all the threads on a node were in a deep sleep state and we spawned
another thread, node->nr_active_cpus was increased, thus a woken-up
thread would not execute the branch after "if (node->nr_deep_sleep-- ==
node->nr_active_cpus)", nr_idle_nodes was not decremented and it would
overflow.

Fix this bug by decrementing nr_idle_nodes in spawn_another_cpu if all
the threads were in the deep-sleep mode and if this is not the first
thread spawned.